[OPS-6138] Always default to the logstash format.
This *can* mean we get two logs for each access, if the vhost logs to
the same file, but at least they will both parse correctly and not
inflate the report of logs hat failed to parse correctly.

In principle, logstash should drop the duplicate identical log message
anyway, as it would generate an identical hash.