[OPS-5712] Configure syslog where it counts, so we get PHP logs in ELK. (#239)
* [OPS-5712] COnfigure syslog where it counts, so we get PHP logs in ELK.

After some fiddling on hidblog, it turns out that the settings in
php-fpm.conf are ignored (or not passed on) and these settings *must*
live under a [syslog] section in a global ini file.

* [OPS-5482] Enable XSS protection via a mapping, for rwint-site compatibility.

* [OPS-5482] Yes, do gzip outgoing SVG files.

* [OPS-6177] Actually,

Why not make the syslog facility tweakable via an overrideable environment
variable, rather than hard-code it? Thank you very much.

* [OPS-6177] Use `local4` as the default syslog facility. ELK knows.

* [HOTFIX] Rather than undoing commits, just pass the hoedown version as an argument.

But, so you even remember how to write Makefiles? You need to add the
arg eh or it all just gets hopelessly ignored.