feat: Allow env-based control of PHP cookie variables.
Default to HTTPONLY=1 and SAMESITE=Strict which will make things more secure (and less Beagle-whiney) unless we specifically override them. This means JS will not be able tom access the session cookie and the session cookie will not be leaked to sites on different domains. In theory this will not break anything.

Refs: OPS-9726